Correção do método que calcula ServicoRealizado.horaConclusao e comentarios

+++ ProjetoBarbearia/src/main/java/br/com/domain/model/entities/Servico.java
@@ -32,5 +32,5 @@ public class Servico {
 	@Column(name = "valor_servico")
 	private BigDecimal valorServico;
 	@Column(name = "duracao_minutos")
-	private Long duracaoMinutos;
+	private Integer duracaoMinutos;
 }
